帖子add，展示

+++ src/main/java/com/libiao/service/impl/AdminServiceImpl.java
@@ -34,7 +34,7 @@ public class AdminServiceImpl implements AdminService {
 
     @Override
     public boolean deletePost(Long id) {
-        if(postService.deletePost(id)){
+        if(postService.deletePost(id.intValue())){
             return true;
         }
         return false;
